+++ dao/src/main/java/be/groups/glanguage/glanguage/api/dao/formula/description/combination/FormulaParameterCombinationDao.java
@@ -10,4 +10,4 @@ import org.springframework.stereotype.*;
 @Repository
 public interface FormulaParameterCombinationDao extends
                                                 JpaRepository<FormulaParameterCombination,
-                                                    Integer> {}
+                                                    Long> {}
