+++ dao/src/main/java/be/groups/glanguage/glanguage/api/dao/rule/definition/RuleDefinitionDao.java
@@ -8,4 +8,4 @@ import org.springframework.stereotype.*;
  * DAO for {@link RuleDefinition} <p>
  */
 @Repository
-public interface RuleDefinitionDao extends JpaRepository<RuleDefinition, Integer> {}
+public interface RuleDefinitionDao extends JpaRepository<RuleDefinition, Long> {}
